Massive updates: - Fix issue with real life board having a delay of half a second to start session - Changeup alot of metrics - Add new plot types

+++ src-tauri/src/types.rs
@@ -62,7 +62,7 @@ pub struct User {
     pub gender: Option<String>,
     pub height: Option<f64>,
     pub height_metric: Option<String>,
-    pub weight: Option<f64>,
+    pub weight: Option<f32>,
     pub weight_metric: Option<String>,
     pub handedness: Option<String>,
     pub color: Option<String>,
